feat(front): make upload a component
