+++ src/pages/upload.tsx
@@ -1,31 +1,5 @@
-export default function Upload() {
-  return (
-    <>
-      <input
-        type="file"
-        accept="image/png, image/jpeg"
-        onChange={(e) => void uploadPhoto(e)}
-      />
-    </>
-  );
+import Upload from "~/components/uploadInput";
+
+export default function UploadPage() {
+  return <Upload />;
 }
-
-const uploadPhoto = async (e: React.ChangeEvent<HTMLInputElement>) => {
-  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion, @typescript-eslint/no-non-null-asserted-optional-chain
-  const file = e.target.files?.[0]!;
-  const filename = file.name;
-  const fileType = file.type;
-
-  const res = await fetch(`/api/upload?file=${filename}&fileType=${fileType}`);
-  const { url } = (await res.json()) as { url: string };
-
-  const uploading = await fetch(url, {
-    method: "PUT",
-    headers: {
-      "Content-Type": fileType,
-    },
-    body: file,
-  });
-
-  console.log(uploading.status);
-};
